requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.24.3 